requirements.txt
yt-dlp==2024.11.18
discord.py==2.4.0
python-dotenv==1.0.1
PyNaCl==1.5.0
discord.py-interactions[voice]==5.13.2